gpg: Configure scdaemon
This is supposed to lower the severity of ykman vs gpg fight.

With the change in place ykman calls work all the time,
but using gpg after ykman was called requires two tries.

First one fails with whatever error it finds pleasing,
but the second one works as expected.

This is to be improved, but still better than the previous
"either ykman works or gpg" requiring manually reinserting the SC.